requirements.txt
python-chess==1.999.0
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1 